requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
aiofiles==23.2.1
aiohappyeyeballs @ file:///C:/b/abs_b505trsapr/croot/aiohappyeyeballs_1725434036096/work
aiohttp @ file:///C:/b/abs_13j6efxjb7/croot/aiohttp_1725529348885/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.6.2.post1
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_c8fgiuixkq/croot/async-timeout_1703097556097/work
attrs @ file:///C:/b/abs_24xj44kobp/croot/attrs_1729089434205/work
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker @ file:///C:/b/abs_d9y2dm7cw2/croot/blinker_1696539752170/work
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools @ file:///C:/b/abs_792zbtc0ua/croot/cachetools_1713977157919/work
catalogue==2.0.10
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
ci_sdr==0.0.2
click @ file:///C:/b/abs_f9ihnt72pu/croot/click_1698129847492/work
cloudpathlib==0.20.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.2.1
coqpit==0.0.17
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
ctc_segmentation==1.7.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
Distance==0.1.3
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
encodec==0.1.1
espnet==202402
espnet-tts-frontend==0.0.3
exceptiongroup==1.2.2
fast_bss_eval==0.1.3
fastapi==0.115.5
ffmpy==0.4.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
Flask==3.1.0
flatbuffers @ file:///C:/b/abs_7f2_85svxz/croot/python-flatbuffers_1722369110965/work
fonttools==4.55.0
frozenlist @ file:///C:/b/abs_06ctmb1zeo/croot/frozenlist_1730903113463/work
fsspec==2024.10.0
g2p-en==2.1.0
g2pkk==0.1.2
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gmpy2 @ file:///C:/ci/gmpy2_1645455782955/work
google-auth @ file:///C:/b/abs_059sl2dhu6/croot/google-auth_1715111206543/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
gradio==4.44.1
gradio_client==1.3.0
grpcio @ file:///C:/b/abs_ccut8qek12/croot/grpc-suite_1681912611694/work
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
h5py @ file:///C:/b/abs_92v137xfqx/croot/h5py_1730215948692/work
hangul-romanize==0.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==4.13.0
importlib_resources==6.4.5
inflect==7.0.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2 @ file:///C:/b/abs_ddyeze2h6_/croot/jinja2_1730903330414/work
joblib==1.4.2
jsonlines==1.2.0
kaldiio==2.18.0
keras @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/keras_1669760570649/work/keras-2.10.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown @ file:///C:/b/abs_98lv_ucina/croot/markdown_1671541919225/work
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib==3.8.4
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
monotonic-align==1.0.0
more-itertools==10.5.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
msgpack==1.1.0
multidict @ file:///C:/b/abs_19e3ubo2ew/croot/multidict_1730905504444/work
murmurhash==1.0.10
mutagen==1.47.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
oauthlib @ file:///C:/b/abs_44riqpe5gg/croot/oauthlib_1679489644705/work
omegaconf==2.3.0
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
orjson==3.10.11
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas==1.5.3
pillow @ file:///C:/b/abs_2cm0qc1vxk/croot/pillow_1731594717982/work
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT @ file:///C:/b/abs_03dm8926m0/croot/pyjwt_1730786800571/work
pynndescent==0.5.13
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.2.0
pypinyin==0.44.0
pyreadline3==3.5.4
pysbd==0.3.4
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
pyworld==0.3.4
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-oauthlib @ file:///C:/b/abs_faw0a7_wag/croot/requests-oauthlib_1720615036938/work
resampy==0.4.3
rich==13.9.4
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
ruff==0.7.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.97
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.0.5
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.3
SudachiDict-core==20241021
SudachiPy==0.6.9
sympy==1.12.1
tensorboard @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorboard_1669760968711/work/tensorboard-2.10.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.10.0
tensorflow-estimator @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorflow-estimator_1669761460695/work/tensorflow_estimator-2.10.0-py2.py3-none-any.whl
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
tomlkit==0.12.0
torch @ file:///C:/b/abs_fakvb73nko/croot/pytorch-select_1730848725921/work
torch-complex==0.4.4
torchaudio==2.3.1
torchvision @ file:///C:/b/abs_36wv6_sd88/croot/torchvision_1722956313919/work
tqdm==4.67.0
trainer==0.0.36
transformers==4.46.3
# Editable install with no version control (TTS==0.22.0)
-e c:\users\amd\tts\tts
typeguard==2.13.3
typer==0.13.1
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.2
tzlocal==5.2
umap-learn==0.5.7
Unidecode==1.3.8
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
uvicorn==0.32.0
wasabi==1.1.3
weasel==0.4.1
websockets==12.0
Werkzeug @ file:///C:/b/abs_b21bhstb1q/croot/werkzeug_1731005800632/work
wget==3.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt @ file:///C:/Windows/Temp/abs_7c3dd407-1390-477a-b542-fd15df6a24085_diwiza/croots/recipe/wrapt_1657814452175/work
yarl @ file:///C:/b/abs_00rcuicx5k/croot/yarl_1726015884215/work
zipp @ file:///C:/b/abs_021kqswk3k/croot/zipp_1729012371206/work
